+++ core/execution/common/interfaces.go
@@ -19,7 +19,7 @@ import (
 
 var One = num.UintOne()
 
-//go:generate go run github.com/golang/mock/mockgen -destination mocks/mocks.go -package mocks code.vegaprotocol.io/vega/core/execution/common TimeService,Assets,StateVarEngine,Collateral,OracleEngine,EpochEngine,AuctionState,LiquidityEngine,EquityLikeShares,MarketLiquidityEngine
+//go:generate go run github.com/golang/mock/mockgen -destination mocks/mocks.go -package mocks code.vegaprotocol.io/vega/core/execution/common TimeService,Assets,StateVarEngine,Collateral,OracleEngine,EpochEngine,AuctionState,LiquidityEngine,EquityLikeShares,MarketLiquidityEngine,Teams,AccountBalanceChecker
 
 // InitialOrderVersion is set on `Version` field for every new order submission read from the network.
 const InitialOrderVersion = 1
@@ -312,4 +312,13 @@ type CommonMarket interface {
 	SubmitStopOrdersWithIDGeneratorAndOrderIDs(context.Context, *types.StopOrdersSubmission, string, IDGenerator, *string, *string) (*types.OrderConfirmation, error)
 
 	PostRestore(context.Context) error
+}
+
+type AccountBalanceChecker interface {
+	GetAvailableBalance(party string) (*num.Uint, error)
+}
+
+type Teams interface {
+	GetTeamMembers(team string, minEpochsInTeam uint64) []string
+	GetAllPartiesInTeams(minEpochsInTeam uint64) []string
 }