+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 - [8969](https://github.com/vegaprotocol/vega/issues/8969) - Improve wiring of internal time triggers for perpetual markets.
 - [9001](https://github.com/vegaprotocol/vega/issues/9001) - Improve wiring of perpetual markets into the data node.
 - [8985](https://github.com/vegaprotocol/vega/issues/8985) - Improve snapshot restore of internal time triggers for perpetual markets.
+- [9146](https://github.com/vegaprotocol/vega/issues/9146) - Improve `TWAP` for perpetual markets to do calculations incrementally.
 - [8817](https://github.com/vegaprotocol/vega/issues/8817) - Add interest term to perpetual funding payment calculation.
 - [8755](https://github.com/vegaprotocol/vega/issues/8755) - Improve testing for perpetual settlement and collateral transfers.
 - [8756](https://github.com/vegaprotocol/vega/issues/8756) - Settlement and margin implementation for `PERPS`.
@@ -65,8 +66,10 @@
 - [9045](https://github.com/vegaprotocol/vega/issues/9045) - Snapshot the referral engine.
 - [9136](https://github.com/vegaprotocol/vega/issues/9136) - Referral engine returns reward factor for a given party
 - [9076](https://github.com/vegaprotocol/vega/issues/9076) - Add current funding rate to market data.
+- [9162](https://github.com/vegaprotocol/vega/issues/9162) - Refactor transfers to support new metric based rewards and distribution strategies
+- [9163](https://github.com/vegaprotocol/vega/issues/9163) - Refactor reward engine to support the new metric based reward distribution
+- [9164](https://github.com/vegaprotocol/vega/issues/9164) - Refactor market activity tracker to support more metrics and history
 - [9219](https://github.com/vegaprotocol/vega/issues/9219) - Don't do `IPFS` garbage collection every segment
-
 - [9208](https://github.com/vegaprotocol/vega/issues/9208) - Refactor referral set and teams state
 - [9204](https://github.com/vegaprotocol/vega/issues/9204) - Ensure teams names are not duplicates